1. 修复RouteTable的问题 2. 删除无用代码 3. 准备使用NIO

+++ src/main/java/me/zzhen/bt/dht/base/Token.java
@@ -14,10 +14,10 @@ import java.time.Instant;
 public class Token {
 
     public final NodeKey key;
-    public final int token;
+    public final long token;
     public final Instant time;
 
-    public Token(NodeKey key, int token, Instant time) {
+    public Token(NodeKey key, long token, Instant time) {
         this.key = key;
         this.token = token;
         this.time = time;
